.gitattributes updated in the correct way, redesign of the roof and new ceiling
The .gitattributes file was modified to be in the correct way for tracking all the files and subfolders of the given path.
What before was called "ceiling" now is treated as the "roof" of the house, and because of that now exists an object that plays the role of a real ceiling instead of being both things.